requirements.txt
Flask==3.1.0
flask_cors==5.0.0
matplotlib==3.9.2
Pillow==11.0.0
ultralytics==8.3.34
waitress==3.0.2
